Shifted NvEncoder class back into its own file
